vector expression release candidate
i rewrote how we parse vector expressions and moved most of the logic over to the tokenizer method. i didn't want to do this but i needed to if i didn't want to clone chaining/method functionality inside of the read_expression of vector expressions. tokenizer is looking pretty busy right now and probably could do with some cleanup. i want to keep this as state-free as possible and as of right now i just introduced the first state the tokenizer has seen so far. #4